adding test_colorCode and moving colorCodingManual into different section
